refactor: add controller
