+++ backend/routes/chat.routes.ts
@@ -6,197 +6,12 @@ import { io } from "..";
 
 const router = Router();
 
-const createRoomBodySchema = z.object({
-  participantAId: z.string(),
-  participantBId: z.string(),
-});
+const {getChat , createChat , updateChat} = require('../controller/chat.control')
 
-router.get("/:id", authenticateToken, async (req, res) => {
-  const { id } = req.params;
+router.get("/:id", authenticateToken, getChat);
 
-  const user = req.body.user;
+router.post("/", authenticateToken, createChat);
 
-  if (id.length != 24 || /[0-9A-Fa-f]{24}/g.test(id) === false) {
-    return res.status(404).send("No chat found");
-  }
-
-  try {
-    // Check if chat is valid
-    const chatRes = await db.chat.findUnique({
-      where: {
-        id: id,
-      },
-      include: {
-        participantA: {
-          select: {
-            id: true,
-            displayName: true,
-            imageURL: true,
-          },
-        },
-        participantB: {
-          select: {
-            id: true,
-            displayName: true,
-            imageURL: true,
-          },
-        },
-        messages: {
-          orderBy: {
-            sentAt: "asc",
-          },
-        },
-      },
-    });
-
-    if (!chatRes) {
-      return res.status(404).send("No chat found");
-    }
-
-    if (
-      chatRes.participantAId !== user.id &&
-      chatRes.participantBId !== user.id
-    ) {
-      return res.status(403).send("You don't have access to this chat");
-    }
-
-    return res.send(chatRes);
-  } catch (err) {
-    return res.status(400).send(err);
-  }
-});
-
-router.post("/", authenticateToken, async (req, res) => {
-  const body = req.body;
-
-  const parseStatus = createRoomBodySchema.safeParse(body);
-  if (!parseStatus.success) return res.status(400).send("Invalid Data");
-
-  const parsedBody = parseStatus.data;
-
-  try {
-    // Check if participant Id is valid or not
-    if (
-      parsedBody.participantAId.length != 24 ||
-      /[0-9A-Fa-f]{24}/g.test(parsedBody.participantAId) === false
-    ) {
-      return res.status(404).send("Participant A not found");
-    }
-
-    if (
-      parsedBody.participantBId.length != 24 ||
-      /[0-9A-Fa-f]{24}/g.test(parsedBody.participantBId) === false
-    ) {
-      return res.status(404).send("Participant B not found");
-    }
-
-    if (parsedBody.participantAId === parsedBody.participantBId) {
-      return res
-        .status(400)
-        .send("You cannot make chat with both of participants are same user");
-    }
-
-    // If there is room already, send that room
-    const findOldChatRoom = await db.chat.findFirst({
-      where: {
-        OR: [
-          {
-            participantAId: parsedBody.participantAId,
-            participantBId: parsedBody.participantBId,
-          },
-          {
-            participantAId: parsedBody.participantBId,
-            participantBId: parsedBody.participantAId,
-          },
-        ],
-      },
-    });
-
-    if (findOldChatRoom) {
-      return res.send(findOldChatRoom);
-    }
-
-    const chatRes = await db.chat.create({
-      data: {
-        participantAId: parsedBody.participantAId,
-        participantBId: parsedBody.participantBId,
-      },
-    });
-
-    io.emit(`users:${parsedBody.participantAId}:chatsUpdate`);
-    io.emit(`users:${parsedBody.participantBId}:chatsUpdate`);
-
-    return res.send(chatRes);
-  } catch (err) {
-    return res.status(400).send(err);
-  }
-});
-
-router.put("/:id/read", authenticateToken, async (req, res) => {
-  const { id } = req.params;
-
-  const user = req.body.user;
-  const userId: string = user.id;
-
-  try {
-    // Check if chat id is valid
-    if (id.length != 24 || /[0-9A-Fa-f]{24}/g.test(id) === false) {
-      return res.status(404).send("Chat not found");
-    }
-
-    // Check if user have priority to read
-    const roomRes = await db.chat.findUnique({
-      where: {
-        id: id,
-        OR: [
-          {
-            participantAId: userId,
-          },
-          {
-            participantBId: userId,
-          },
-        ],
-      },
-    });
-
-    if (!roomRes) {
-      return res
-        .status(403)
-        .send("You don't have priority to update this chat");
-    }
-
-    if (roomRes.participantAId === userId) {
-      // Update A
-      const updateRes = await db.chat.update({
-        where: {
-          id: id,
-        },
-        data: {
-          participantAUnread: 0,
-        },
-      });
-
-      io.emit(`users:${userId}:chatsUpdate`);
-
-      return res.send(updateRes);
-    } else {
-      // Update B
-      const updateRes = await db.chat.update({
-        where: {
-          id: id,
-        },
-        data: {
-          participantBUnread: 0,
-        },
-      });
-
-      io.emit(`users:${userId}:chatsUpdate`);
-
-      return res.send(updateRes);
-    }
-  } catch (err) {
-    return res.status(403).send(err);
-  }
-});
+router.put("/:id/read", authenticateToken, updateChat);
 
 export default router;
